+++ R/main.r
@@ -36,5 +36,7 @@
   invisible()
 }
 
+# utils::globalVariables(c("scale_tabs"))
+