+++ R/schedenveg.r
@@ -6,4 +6,8 @@
 #'
 #' @format A data frame with 28 rows (samples) and 155 variables (species)
 #' @references Eichholz, A. (1997): Wiesen und Magerrasen am Suedhang des Hohen Hagen. Diplomarbeit Biologie, University of Goettingen.
+#' @docType data
+#' @keywords datasets
+#' @name schedenveg
+#' @usage data(schedenveg)
 "schedenveg"